Rename DeployQueries to DeployQuery
